requirements.txt
pandas>=2.0.0
google-generativeai>=0.3.0
